+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Added Stringable to interface
+* Added Stringable suport to interface
+* Detect 16 byte strings in fromString()
 
 ## v0.1.2 (2023-11-17)
 * Added create*String() methods